Move go_suffix_for_arch() to artcommon
