+++ doozer/doozerlib/cli/release_gen_payload.py
@@ -12,13 +12,13 @@ from unittest.mock import MagicMock
 import aiofiles
 import click
 
-from artcommonlib.arch_util import brew_arch_for_go_arch
+from artcommonlib.arch_util import brew_arch_for_go_arch, go_suffix_for_arch
 from doozerlib import rpm_utils
 import yaml
 from artcommonlib import rhcos
 from artcommonlib.rhcos import RhcosMissingContainerException
 import openshift as oc
-from opentelemetry import metrics, trace
+from opentelemetry import trace
 
 
 from doozerlib.rpm_utils import parse_nvr
@@ -29,8 +29,7 @@ from doozerlib.image import ImageMetadata, BrewBuildImageInspector, ArchiveImage
 from doozerlib.assembly_inspector import AssemblyInspector
 from doozerlib.runtime import Runtime
 from doozerlib.telemetry import start_as_current_span_async
-from doozerlib.util import red_print, go_suffix_for_arch, isolate_nightly_name_components, \
-    go_arch_for_brew_arch
+from doozerlib.util import red_print, isolate_nightly_name_components, go_arch_for_brew_arch
 from artcommonlib.util import convert_remote_git_to_https
 from doozerlib.assembly import AssemblyTypes, assembly_basis, AssemblyIssue, AssemblyIssueCode
 from doozerlib import exectools
