+++ pyartcd/pyartcd/pipelines/promote.py
@@ -21,10 +21,10 @@ from tenacity import (RetryCallState, RetryError, retry,
                       retry_if_exception_type, retry_if_result,
                       stop_after_attempt, wait_fixed)
 
-from artcommonlib.arch_util import brew_suffix_for_arch, brew_arch_for_go_arch
+from artcommonlib.arch_util import brew_suffix_for_arch, brew_arch_for_go_arch, go_suffix_for_arch
 from artcommonlib.rhcos import get_primary_container_name
 from doozerlib import assembly
-from doozerlib.util import (go_arch_for_brew_arch, go_suffix_for_arch)
+from doozerlib.util import go_arch_for_brew_arch
 from pyartcd.locks import Lock
 from pyartcd.signatory import AsyncSignatory
 from pyartcd.util import nightlies_with_pullspecs
